fix bug

+++ src/cpu/instruction.c
@@ -356,7 +356,7 @@ int cpu_step(CPU *cpu) {
       break;
     }
     case OP_ASL: {
-      uint8_t addr = cpu_fetch_operand(cpu, inst.a);
+      uint16_t addr = cpu_fetch_operand(cpu, inst.a);
       uint8_t m = inst.a == ADDR_ACCUMULATOR ?
         cpu->reg.a :
         cpubus_read(cpu->bus, addr);
@@ -374,7 +374,7 @@ int cpu_step(CPU *cpu) {
       break;
     }
     case OP_LSR: {
-      uint8_t addr = cpu_fetch_operand(cpu, inst.a);
+      uint16_t addr = cpu_fetch_operand(cpu, inst.a);
       uint8_t m = inst.a == ADDR_ACCUMULATOR ?
         cpu->reg.a :
         cpubus_read(cpu->bus, addr);
@@ -394,7 +394,7 @@ int cpu_step(CPU *cpu) {
     case OP_ROL: {
       uint8_t carry = cpu_get_pflag(cpu, P_STATUS_CARRY);
 
-      uint8_t addr = cpu_fetch_operand(cpu, inst.a);
+      uint16_t addr = cpu_fetch_operand(cpu, inst.a);
       uint8_t m = inst.a == ADDR_ACCUMULATOR ?
         cpu->reg.a :
         cpubus_read(cpu->bus, addr);
@@ -414,7 +414,7 @@ int cpu_step(CPU *cpu) {
     case OP_ROR: {
       uint8_t carry = cpu_get_pflag(cpu, P_STATUS_CARRY);
 
-      uint8_t addr = cpu_fetch_operand(cpu, inst.a);
+      uint16_t addr = cpu_fetch_operand(cpu, inst.a);
       uint8_t m = inst.a == ADDR_ACCUMULATOR ?
         cpu->reg.a :
         cpubus_read(cpu->bus, addr);
@@ -432,7 +432,7 @@ int cpu_step(CPU *cpu) {
       break;
     }
     case OP_INC: {
-      uint8_t addr = cpu_fetch_operand(cpu, inst.a);
+      uint16_t addr = cpu_fetch_operand(cpu, inst.a);
       uint8_t res = cpubus_read(cpu->bus, addr) + 1;
       cpubus_write(cpu->bus, addr, res);
 
@@ -458,7 +458,7 @@ int cpu_step(CPU *cpu) {
       break;
     }
     case OP_DEC: {
-      uint8_t addr = cpu_fetch_operand(cpu, inst.a);
+      uint16_t addr = cpu_fetch_operand(cpu, inst.a);
       uint8_t res = cpubus_read(cpu->bus, addr) - 1;
       cpubus_write(cpu->bus, addr, res);
